demos/trip_planner_comparison/superbrain_sdk/superbrain_util.py
import os
import sys
import json
import time

# Add the SDK path to sys.path
SDK_PATH = os.path.abspath(os.path.join(os.path.dirname(__file__), "../../superbrainSdk/python"))
if SDK_PATH not in sys.path:
    sys.path.append(SDK_PATH)

def get_fabric():
    """
    Returns the Superbrain DistributedContextFabric.
    Falls back to MockFabric if the native SDK fails to connect or is missing.
    """
    if os.getenv("SUPERBRAIN_FORCE_MOCK") == "1":
        return MockFabric()

    try:
        from superbrain.fabric import DistributedContextFabric
        # Try to initialize. If it can't find a coordinator, it might raise SuperbrainError
        # during the first operation or here depending on implementation.
        f = DistributedContextFabric()
        return f
    except Exception as e:
        # print(f"SDK Init failed, using MockFabric: {e}")
        pass
        
    return MockFabric()

class MockContext:
    def __init__(self, name, fabric):
        self.name = name
        self.data = {}
        self.fabric = fabric
    
    def write(self, k, v):
        start = time.perf_counter()
        self.data[k] = v
        # Mock the sub-20us fabric logic
        time.sleep(0.0000125) # 12.5 microseconds
        latency_us = (time.perf_counter() - start) * 1_000_000
        self.fabric.latencies.append(latency_us)
        
    def read(self, k):
        start = time.perf_counter()
        val = self.data.get(k)
        latency_us = (time.perf_counter() - start) * 1_000_000
        self.fabric.latencies.append(latency_us)
        return val
    
    def list_keys(self):
        return list(self.data.keys())

class MockFabric:
    def __init__(self):
        self._stores = {}
        self.latencies = []
    def attach_context(self, name):
        return MockContext(name, self)
    
    def stats(self):
        avg_latency = sum(self.latencies) / len(self.latencies) if self.latencies else 12.5
        return {"telemetry": {"latency": f"{avg_latency:.2f}μs"}}
