demos/trip_planner_comparison/superbrain_sdk/pyproject.toml
[project]
name = "trip-planner-crew"
version = "0.1.0"
description = ""
authors = [{name = "Your Name", email = "you@example.com"}]
requires-python = ">=3.10.0,<3.12"
dependencies = [
    "crewai>=0.30.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.0.10",
    "unstructured>=0.14.3",
    "pyowm==3.3.0",
    "python-dotenv==1.0.0",
]

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

